Integrate animal sniffer
Animal sniffer provides tools to assist verifying that classes
compiled with a newer JDK are compatible with an older JDK.

This integratesthe latest version of the tool (1.11) that
permits the use of custom annotations. Added @Java7Required,
@Java8Required and @SunVmRequired and annotated the few places
where we rely on a specific environment.

The verification process can be invoked by running the 'sniff'
task.