JMS annotation-driven endpoints.
This commit adds a java config support for JMS annotated endpoint
defined by the @JmsListener interface. Can be activated for now by
the @EnableJms on a configuration class.

Containers are created and managed under the cover by a registry
at application startup time. Container creation is delegated to a
set of JmsListenerContainerFactory that are identified by a factoryId
present on the JmsListener annotation.

The configuration can be fine-tuned by implementing the
JmsListenerConfigurer interface and get access to the registrar used
to register endpoints. This includes a programmatic registration of
endpoints in complement to the declarative approach.