Add easier overriding of CacheInterceptor
Prior to this commit, internal CacheInterceptor and JCacheInterceptor
instances were not configured consistently: java config support exposed
a bean with name 'cacheInterceptor' while the XML support exposed a
bean with a generated name.

This commit harmonizes the whole infrastructure, including related
components such as CacheOperationSource.

This make the overriding of CacheInterceptor consistent. Added a
invokeOperation protected method in case one needs a hook point in
the way the underlying cache method is invoked, and how exceptions
that might be thrown by that invocation are handled.

Issue: SPR-11540