Add lifecycle option for EhCache and JCache managers
Both EhCache and JCache support the creation of a dedicated cache
infrastructure or the reuse of an existing one. The factory bean counter
part provides a shutdown lifecycle that is not really applied with Java
config since these factory beans aren't generally used.

Update EhCacheCacheManager and JCacheCacheManager to expose dedicated
methods to define if the underlying cache manager should be closed on
shutdown of the context.

Issue: SPR-12841