Support for Spring components indexer
This commit adds a "spring-context-indexer" module to Spring Framework
that can be added to any project in order to generate an index of Spring
components defined in the project.

`SpringComponentsIndexer` is a standard annotation processor that looks
for source files with target annotations (typically `@Component`) and
references them in a `META-INF/spring.components` generated file.

Such files can later be reused  by the `ApplicationContext` to avoid
using component scan. A global `SpringComponentsIndex` can be easily
loaded from the current classpath using `SpringComponentsIndexLoader`.

Issue: SPR-11890