Provide an API to record various runtime hints
This commit provides an API to record the need for reflection,
resources, serialization, and proxies at runtime. It offers a central
`RuntimeHints` facility that can record various aspects.

For reflection, individual elements of a type can be recorded, as well
as a predefined categories (identified by the `MemberCategory` enum). A
method or constructor hint can refine whether the need to execute it is
required.

For resources, they can be included using includes/excludes pattern that
can either:

* An exact match: `com/example/test.properties`
* A pattern, using `*` or `?` to include multiple resources, such as
`com/example/*.properties`
* A regular expression

Resource bundles can be added as well by providing its basename.

// TODO serialization - proxy

Closes gh-27829