Update ClassNameGenerator to expose a configurable target resolver
This commit updates ClassNameGenerator so that it can resolve the
chosen target for a feature. Rather than using the target itself, a
function can determine where the code should be written, based on the
candidate target.

This commit also provides an implementation of such resolver that adds
a configurable package name prefix. This lets generated code to be put
under a configurable package, rather than being written in the raw
target. Doing so can lead to split package that certain environments
do not tolerate.