Invoke ConfigurationBeanNameGenerator even if a name is provided
This commit updates ConfigurationBeanNameGenerator so that it is
invoked with the chosen bean name, as specified on `@Bean`. If no
such bean is present, `null` is provided. The default implementation
returns the bean name if provided, but other implementations can now
decide to override the provided bean name if necessary.

Closes gh-35505